+++ components/input/index.js
@@ -1,2 +1,6 @@
 import Input from './Input';
+import InputBase from './InputBase';
+import NumericInput from './NumericInput';
+
+export { InputBase, NumericInput };
 export default Input;